fix(refs DPLAN-11516): move to mixin
