+++ client/js/components/procedure/SegmentsList/SegmentsList.vue
@@ -246,14 +246,10 @@
         </div>
       </template>
 
-      <div v-else>
-        <p class="flash flash-info">
-          <i
-            class="fa fa-info-circle"
-            aria-hidden="true" />
-          {{ Translator.trans('segments.none') }}
-        </p>
-      </div>
+      <dp-inline-notification
+        v-else
+        :message="Translator.trans('segments.none')"
+        type="info" />
     </template>
   </div>
 </template>
@@ -268,6 +264,7 @@ import {
   DpColumnSelector,
   DpDataTable,
   DpFlyout,
+  DpInlineNotification,
   DpLoading,
   DpPager,
   dpRpc,
@@ -283,6 +280,7 @@ import fullscreenModeMixin from '@DpJs/components/shared/mixins/fullscreenModeMi
 import lscache from 'lscache'
 import paginationMixin from '@DpJs/components/shared/mixins/paginationMixin'
 import StatementMetaTooltip from '@DpJs/components/statement/StatementMetaTooltip'
+import tableScrollbarMixin from '@DpJs/components/shared/mixins/tableScrollbarMixin'
 
 export default {
   name: 'SegmentsList',
@@ -294,6 +292,7 @@ export default {
     DpColumnSelector,
     DpDataTable,
     DpFlyout,
+    DpInlineNotification,
     DpLoading,
     DpPager,
     DpStickyElement,
@@ -306,7 +305,7 @@ export default {
     cleanhtml: CleanHtml
   },
 
-  mixins: [fullscreenModeMixin, paginationMixin, tableSelectAllItems],
+  mixins: [fullscreenModeMixin, paginationMixin, tableScrollbarMixin, tableSelectAllItems],
 
   props: {
     currentUserId: {
@@ -557,21 +556,6 @@ export default {
         })
     },
 
-    /**
-     * Adjust the width of the inner element of the footer scrollbar to the width of the Table.
-     */
-    updateScrollbarStyles () {
-      const tableWidth = window.getComputedStyle(this.dataTableElement).width
-      const tableContainerWidth = window.getComputedStyle(this.dataTableContainerElement).width
-
-      if (tableWidth > tableContainerWidth) {
-        this.scrollbar.classList.remove('hidden')
-        this.scrollbar.firstChild.setAttribute('style', 'width:' + tableWidth + ';height:1px;')
-      } else {
-        this.scrollbar.classList.add('hidden')
-      }
-    },
-
     fetchSegmentIds (payload) {
       return dpRpc('segment.load.id', payload)
         .then(response => checkResponse(response))
@@ -703,36 +687,6 @@ export default {
       this.resetSelection()
       this.applyQuery(1)
     }
-  },
-
-  created () {
-    this.$watch('isLoading', (isLoading) => {
-      if (isLoading) {
-        return
-      }
-
-      this.$nextTick(() => {
-        this.scrollbar = this.$refs.scrollBar
-        this.dataTableContainerElement = this.$refs.dataTable.$el
-        this.dataTableElement = this.$refs.dataTable.$refs.tableEl
-
-        // Bind behaviour and position of the footer scrollbar to the scroll position of the dataTableContainerElement.
-        this.scrollbar.addEventListener('scroll', () => {
-          this.dataTableContainerElement.scrollLeft = this.scrollbar.scrollLeft
-        })
-
-        this.dataTableContainerElement.addEventListener('scroll', () => {
-          this.scrollbar.scrollLeft = this.dataTableContainerElement.scrollLeft
-        })
-
-        // Observe changes to dataTable to update scrollbar accordingly
-        this.dataTableObserver = new ResizeObserver(this.updateScrollbarStyles.bind(this))
-        this.dataTableObserver.observe(this.dataTableElement)
-
-        // Set scrollbar width or conditionally hide it.
-        this.updateScrollbarStyles()
-      })
-    })
   },
 
   mounted () {
